Address Rack Builder review feedback
