+++ src/pages/rack-builder/PanelLayoutsOverviewPage.tsx
@@ -1,19 +1,14 @@
-import { useNavigate, useSearchParams } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
+import { useActiveRackBuilderDepartment } from '@/features/rack-builder/hooks/useActiveRackBuilderDepartment'
 import { useProjects } from '@/features/rack-builder/hooks/useProjects'
 import { usePanelLayoutCounts } from '@/features/rack-builder/hooks/usePanelLayoutCounts'
-import { useOptimizedAuth } from '@/hooks/useOptimizedAuth'
-import { getRackBuilderProjectsPath, normalizeRackBuilderDepartment } from '@/features/rack-builder/lib/department'
+import { getRackBuilderProjectsPath } from '@/features/rack-builder/lib/department'
 import PageHeader from '@/features/rack-builder/components/layout/PageHeader'
 import Button from '@/features/rack-builder/components/ui/Button'
 
 export default function PanelLayoutsOverviewPage() {
   const navigate = useNavigate()
-  const [searchParams] = useSearchParams()
-  const { userDepartment } = useOptimizedAuth()
-  const activeDepartment =
-    normalizeRackBuilderDepartment(searchParams.get('department'))
-    ?? normalizeRackBuilderDepartment(userDepartment)
-    ?? 'sound'
+  const activeDepartment = useActiveRackBuilderDepartment()
   const { projects, loading: projectsLoading } = useProjects(activeDepartment)
   const { counts, loading: countsLoading } = usePanelLayoutCounts()
